update quadratic solve

+++ quadratic-solver.py
@@ -1,6 +1,6 @@
-a = int(input('Enter a: '))
-b = int(input('Enter b: '))
-c = int(input('Enter c: '))
+a = float(input('Enter a: '))
+b = float(input('Enter b: '))
+c = float(input('Enter c: '))
 d = (b**2 - 4*a*c)**(1/2)
 
 x1 = (-b + d)/2
